Fix unsorted full off-by-one
Finally get unsorted full conditions to actually have proper semantics.
This also formalizes the behavior, which is that the first item to put
the unsorted size over the threshhold will cause it to be made into a
block.

Also substantially strengthen the tests of block merging and arbitary
full predicates.