Add basic "indexed" block store
Can ingress blocks and merge them with blocks currently in the index.
This is conceptually a two level structure: the top level is the "keys"
(each block minimum is the key for that block in the ordered-map) and
the bottom level is the block itself, which contains a sequence of
items.

A structure with indefinite levels would be something like a btree.
This is simpler and can always be replaced with a btree if scale needs
end up requiring it.

The min key and max key are fetched directly from the block, but the
actual ordered items in the block are not used. This means the block
could switch to a different internal representation relatively easily
(though a protocol for min and max would need to be designed and
implemented).