Add basic building block to query indexed store
blocks-overlapping-range looks for blocks which could be the basis of a
query. As long as the range being looked for is near the "high end" of
the indexed store or the blocks are sufficiently large, this is
reasonably efficient. However, due to the two-level nature of the store,
higher performance will be elusive.