Allow order-blocks/items to return any seq(able)
Requiring a vector is confusing and restrictive. Very few locations
actually depend on the return type, and they can just coerce it to a
vector.

Note that the current impl still returns a vector, since this is how it
is stored internally.

TODO: create a test impl that returns it as a seq and verify that the
tests pass. This works manually, but some additional work is required to
write a test for it.