Append returns new unsorted rather than only vec
Previously, it just appended to a vector, requring callers to know about
the interal structure of the unsorted to get the vector out and put it
back in. This made a bit more sense when the trans-store and unsorted
were more tightly coupled, but now that they are being decoupled it
makes sense for the append operation to be a more conventional
functional operation. This will make more sense once "empty out all the
items" and "get the items" have APIs as well, rather than requiring
reaching into the map directly.